requirements.txt
fastapi
uvicorn
sqlalchemy
torch
diffusers
Pillow
